Organize unit tests into relevant folder
